Adicionar estrutura para API, incluir rotas para API e ordenar dados de forma alfabetica
