+++ app/models/user.rb
@@ -3,4 +3,6 @@ class User < ApplicationRecord
     validates :email, presence: true
     validates :password, presence: true
     has_many :stores
+
+    scope :order_user_by_name, -> { order(:name) }
 end
